moved the deleteBranchesFilesAndCommits method to the DeleteProjectPort
