+++ coderadar-core/src/main/java/io/reflectoring/coderadar/projectadministration/port/driven/project/DeleteProjectPort.java
@@ -8,4 +8,11 @@ public interface DeleteProjectPort {
    * @param projectId The id of the project to delete.
    */
   void delete(long projectId);
+
+  /**
+   * Deletes all files and commits for the given project
+   *
+   * @param projectId The id of the project.
+   */
+  void deleteBranchesFilesAndCommits(long projectId);
 }
